Fix compilation error on arm
Summary:
We build on 64 bit platforms where 0L is a 64 bit integer, but that
isn't the case on our arm android builds. Thus this was complaining
during compilation that there is not max(int32, int64), only template<T>
max(T& lhs, T& rhs);

Test Plan: compiled

Reviewed By: davejwatson@fb.com

Blame Rev:

+++ thrift/lib/cpp/async/TAsyncServerSocket.cpp
@@ -1,21 +1,19 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
+ * Copyright 2014 Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 #define __STDC_FORMAT_MACROS
 
 #include "thrift/lib/cpp/async/TAsyncServerSocket.h"
@@ -597,7 +595,8 @@ void TAsyncServerSocket::handlerReady(
     address.addressUpdated(addressFamily, addrLen);
 
     int64_t nowMs = concurrency::Util::currentTime();
-    int64_t timeSinceLastAccept = std::max(0L, nowMs - lastAccepTimestamp_);
+    int64_t timeSinceLastAccept = std::max(int64_t(0),
+        nowMs - lastAccepTimestamp_);
     lastAccepTimestamp_ = nowMs;
     if (acceptRate_ < 1) {
       acceptRate_ *= 1 + acceptRateAdjustSpeed_ * timeSinceLastAccept;
